requirements.txt
openai
   langchain
   langchain-community
   python-dotenv

fastapi
uvicorn
apscheduler
geopy
